requirements.txt
aiofiles==23.1.0
anyio==3.6.2
certifi==2022.12.7
click==8.1.3
colorama==0.4.6
exceptiongroup==1.1.1
fastapi==0.95.1
flake8==5.0.4
h11==0.14.0
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
idna==3.4
importlib-metadata==4.2.0
iniconfig==2.0.0
isort==5.0.0
mccabe==0.7.0
packaging==23.1
pluggy==1.0.0
pycodestyle==2.9.1
pydantic==1.10.7
pyflakes==2.5.0
pytest==7.3.1
python-dotenv==0.21.1
PyYAML==6.0
sniffio==1.3.0
starlette==0.26.1
toml==0.10.2
tomli==2.0.1
typing_extensions==4.5.0
uvicorn==0.21.1
watchfiles==0.19.0
websockets==11.0.1
zipp==3.15.0
